refactor(inference): isolate featured model catalog
